+++ mandelbulber2/qt/dock_navigation.cpp
@@ -83,7 +83,7 @@ void cDockNavigation::SetIconSizes()
 {
 	QFontMetrics fm(font());
 	int pixelFontSize = fm.height();
-	QSize iconSize(pixelFontSize * 2.5, pixelFontSize * 2.5);
+	QSize iconSize(int(pixelFontSize * 2.5), int(pixelFontSize * 2.5));
 	ui->bu_move_backward->setIconSize(iconSize);
 	ui->bu_move_down->setIconSize(iconSize);
 	ui->bu_move_forward->setIconSize(iconSize);
